+++ server/TWS Foundation/TWS Customer.Quality/Q_Features/Q_Bussines/Q_LoadTypesService.cs
@@ -5,6 +5,7 @@ using CSM_Foundation.Database.Entity.Models.Output;
 
 using TWS_Business.Depots;
 using TWS_Business.Depots.Vehicles;
+using TWS_Business.Entities;
 using TWS_Business.Entities.Vehicules;
 
 using TWS_Customer.Features.Business.Vehicules;
@@ -101,6 +102,19 @@ public class Q_LoadTypesService
            () => Assert.False(batchOutput.Failed),
            () => Assert.Equal(3, batchOutput.Successes.Length),
            () => Assert.Empty(batchOutput.Failures)
+        );
+    }
+
+    [Fact(DisplayName = "[Read]: Read a record filtering by reference property.")]
+    public async Task Read() {
+        // Create a sample to prevent empty read results.
+        LoadType load = await _depot!.Store(SampleLoadtype(), true);
+        BatchOperationOutput<LoadType> batchResult = await _service.Read(load.Reference);
+
+        Assert.Multiple(
+            () => Assert.True(batchResult.SuccessesCount > 0),
+            () => Assert.Equal(0, batchResult.FailuresCount),
+            () => Assert.Single(batchResult.Successes)
         );
     }
 }